# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис сокращения ссылок с поддержкой сессий, батчинга и асинхронного удаления.",
        "title": "MicroURL Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/": {
            "post": {
                "description": "Принимает длинный URL в теле запроса как строку и возвращает короткую ссылку.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "URL"
                ],
                "summary": "Сократить URL (Text)",
                "parameters": [
                    {
                        "description": "Оригинальный длинный URL",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Короткая ссылка",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "URL уже сокращен (возвращает существующую ссылку)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Домен заблокирован (циклическое сокращение)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/shorten": {
            "post": {
                "description": "Принимает JSON {\"url\": \"...\"} и возвращает JSON {\"result\": \"...\"}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL API"
                ],
                "summary": "Сократить URL (JSON)",
                "parameters": [
                    {
                        "description": "Запрос сокращения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.APIShortenReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешное создание",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIShortenResp"
                        }
                    },
                    "400": {
                        "description": "Неверный формат JSON или URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Конфликт: уже существует",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIShortenResp"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/shorten/batch": {
            "post": {
                "description": "Принимает массив объектов и возвращает массив сокращенных ссылок. Поддерживает большие объемы данных.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL API"
                ],
                "summary": "Пакетное сокращение (Streaming)",
                "parameters": [
                    {
                        "description": "Массив ссылок для сокращения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BatchItem"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Массив результатов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BatchItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный формат",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/urls": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список всех когда-либо сокращенных пользователем ссылок. Требуется авторизация (Cookie или Bearer).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Ссылки пользователя",
                "responses": {
                    "200": {
                        "description": "Список ссылок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.URLItem"
                            }
                        }
                    },
                    "204": {
                        "description": "У пользователя нет ссылок",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Принимает список ID ссылок и асинхронно помечает их как удаленные. Только для владельца.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Удалить ссылки (Batch)",
                "parameters": [
                    {
                        "description": "Массив ID коротких ссылок",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Принято к обработке",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Проверка доступности соединения с хранилищем.",
                "tags": [
                    "System"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "DB unreachable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}": {
            "get": {
                "description": "Возвращает 307 статус и заголовок Location для перехода.",
                "tags": [
                    "Redirect"
                ],
                "summary": "Переход по ссылке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID короткой ссылки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Redirect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "410": {
                        "description": "Ссылка удалена",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BatchItem": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "description": "CorrelationID — произвольный идентификатор строки, предоставленный клиентом.\nСлужит для сопоставления запроса и ответа в асинхронных или пакетных операциях.",
                    "type": "string"
                },
                "err": {
                    "description": "Err содержит описание ошибки обработки конкретного элемента, если она возникла.",
                    "type": "string"
                },
                "original_url": {
                    "description": "OriginalURL — исходный длинный URL.\nМожет быть пустым в ответе, если клиент уже знает его.",
                    "type": "string"
                },
                "short_url": {
                    "description": "ShortURL — результат сокращения для данной строки.",
                    "type": "string"
                }
            }
        },
        "dto.URLItem": {
            "type": "object",
            "properties": {
                "correlation_id": {
                    "description": "CorrelationID — идентификатор, связывающий запись с исходным запросом (если применимо).",
                    "type": "string"
                },
                "original_url": {
                    "description": "OriginalURL — исходный длинный URL.",
                    "type": "string"
                },
                "short_url": {
                    "description": "ShortURL — соответствующий короткий URL.",
                    "type": "string"
                }
            }
        },
        "handlers.APIShortenReq": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "handlers.APIShortenResp": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Тип 'Bearer \u003ctoken\u003e' или использование куки 'murl_session'",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}